% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmaxno.R
\name{getmaxno}
\alias{getmaxno}
\title{getmaxno 用來找stata 輸出的檔案,利用指定的型態,
在chunk 執行之前,找出目前的最大號碼,例如3,在chunk 執行之後
如果發現大於3的號碼,都會被認為是stata chunk產生的檔案,然後連結.}
\usage{
getmaxno(pattern, ext, adir = "report/")
}
\description{
getmaxno 用來找stata 輸出的檔案,利用指定的型態,
在chunk 執行之前,找出目前的最大號碼,例如3,在chunk 執行之後
如果發現大於3的號碼,都會被認為是stata chunk產生的檔案,然後連結.
}
