% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html_dependencies}
\alias{html_dependencies}
\alias{html_dependency_clipboard}
\alias{html_dependency_allcodecopy}
\title{Provide HTML dependencies}
\usage{
html_dependency_clipboard()

html_dependency_allcodecopy()
}
\value{
An object that can be included in a list of dependencies passed to
\link[htmltools:htmlDependencies]{htmltools::attachDependencies()}.
}
\description{
These functions provide HTML dependencies for \code{clipboard.js},
}
\seealso{
Other HTML dependencies functions: 
\code{\link{SmdAid_dependencies}()}
}
\concept{HTML dependencies functions}
