% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd_html.R, R/smd_html2.R
\name{smd_html}
\alias{smd_html}
\title{my knit.}
\usage{
smd_html(..., output_dir = "report")

smd_html(..., output_dir = "report")
}
\description{
render to sub folder report

render to sub folder report
}
