% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{SmdAid_dependencies}
\alias{SmdAid_dependencies}
\title{List SmdAid dependencies}
\usage{
SmdAid_dependencies()
}
\value{
A list of dependencies that can be passed to
\link[htmltools:htmlDependencies]{htmltools::attachDependencies()} or
\code{\link[rmarkdown:html_document_base]{rmarkdown::html_document_base()}}.
}
\description{
This function is used to get the list of \code{SmdAid} dependencies.
}
\seealso{
Other HTML dependencies functions: 
\code{\link{html_dependencies}}
}
\concept{HTML dependencies functions}
